+++ package/testTree.js
@@ -1,0 +1,1 @@
+import treeGraphFactory from './treeGraphFactory';
